Added a graph building and visualization script for a toy word-pair grammar
The grammar in fact was changed to use words with ambiguity in
the lexicon to illustrate the use of auxiliary symbols. In fact some
scripts seem to expect that such symbols are present.
A minor change was made in scripts/make_phones_symtab.pl . As it was
it was reading the pronunciations starting from the second phone.
I don't know if this is a bug or handles some special cases, while relying
on bigger(i.e. non-toy) lexicons.

The graphs built are stored in data_toy and the pictures in
data_toy/pictures.